# backend/src/test/resources/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 공통기능 25개 API
  version: 0.1.0
servers:
- url: /
paths:
...
  /api/admin/menu-permissions:
    get:
      operationId: listMenuPermissions
      summary: 메뉴 권한 관리 조회
      x-related-requirements:
      - REQ-021
      - REQ-022
      - REQ-023
      - REQ-024
      - REQ-105
      - REQ-107
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-119
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-142
      - REQ-166
      - REQ-167
      - REQ-174
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      - name: filter
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveMenuPermissions
      summary: 메뉴 권한 관리 저장/실행
      x-related-requirements:
      - REQ-021
      - REQ-022
      - REQ-023
      - REQ-024
      - REQ-105
      - REQ-107
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-119
      - REQ-120
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-142
      - REQ-166
      - REQ-167
      - REQ-174
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      x-business-rules:
      - 메뉴 권한 관리 원문 제약과 서버 권한을 적용한다
      - menu_permissions 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - menu_permissions 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 메뉴 권한 관리 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 메뉴 권한 관리 저장 -> menu_permissions 후속 조회에 반영'
      - 'auth: R01이 메뉴 권한 관리 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> menu_permissions 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 메뉴 권한 관리 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
  /api/admin/system-configurations:
    get:
      operationId: getSystemConfigurations
      summary: 공통 환경설정 조회
      x-related-requirements:
      - REQ-042
      - REQ-043
      - REQ-044
      - REQ-105
      - REQ-107
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-119
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-148
      - REQ-166
      - REQ-167
      - REQ-180
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveSystemConfigurations
      summary: 공통 환경설정 저장/실행
      x-related-requirements:
      - REQ-042
      - REQ-043
      - REQ-044
      - REQ-105
      - REQ-107
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-119
      - REQ-120
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-148
      - REQ-166
      - REQ-167
      - REQ-180
      - REQ-194
      - REQ-198
      - REQ-218
      x-roles:
      - R09
      x-business-rules:
      - 공통 환경설정 원문 제약과 서버 권한을 적용한다
      - system_configurations 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - system_configurations 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 공통 환경설정 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 공통 환경설정 저장 -> system_configurations 후속 조회에 반영'
      - 'auth: R01이 공통 환경설정 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> system_configurations 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 공통 환경설정 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
  /api/admin/notices:
    get:
      operationId: listNotices
      summary: 공지사항 관리 조회
      x-related-requirements:
      - REQ-052
      - REQ-053
      - REQ-054
      - REQ-055
      - REQ-105
      - REQ-108
      - REQ-109
      - REQ-112
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-151
      - REQ-166
      - REQ-167
      - REQ-183
      - REQ-194
      - REQ-198
      - REQ-202
      - REQ-204
      - REQ-212
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      - name: filter
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: saveNotices
      summary: 공지사항 관리 저장/실행
      x-related-requirements:
      - REQ-052
      - REQ-053
      - REQ-054
      - REQ-055
      - REQ-105
      - REQ-108
      - REQ-109
      - REQ-112
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-120
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-151
      - REQ-166
      - REQ-167
      - REQ-183
      - REQ-194
      - REQ-198
      - REQ-202
      - REQ-204
      - REQ-212
      x-roles:
      - R09
      x-business-rules:
      - 공지사항 관리 원문 제약과 서버 권한을 적용한다
      - notices 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - notices 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 공지사항 관리 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 공지사항 관리 저장 -> notices 후속 조회에 반영'
      - 'auth: R01이 공지사항 관리 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> notices 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 공지사항 관리 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
  /api/admin/privacy-policies:
    get:
      operationId: listPrivacyPolicies
      summary: 개인정보 관리 조회
      x-related-requirements:
      - REQ-076
      - REQ-077
      - REQ-078
      - REQ-079
      - REQ-080
      - REQ-081
      - REQ-105
      - REQ-111
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-123
      - REQ-124
      - REQ-126
      - REQ-156
      - REQ-157
      - REQ-158
      - REQ-166
      - REQ-167
      - REQ-188
      - REQ-194
      - REQ-198
      - REQ-211
      x-roles:
      - R09
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        schema:
          type: integer
          enum:
          - 20
          - 50
          - 100
          default: 20
      - name: sort
        in: query
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      - name: filter
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
    post:
      operationId: savePrivacyPolicies
      summary: 개인정보 관리 저장/실행
      x-related-requirements:
      - REQ-076
      - REQ-077
      - REQ-078
      - REQ-079
      - REQ-080
      - REQ-081
      - REQ-105
      - REQ-111
      - REQ-114
      - REQ-115
      - REQ-116
      - REQ-117
      - REQ-120
      - REQ-126
      - REQ-133
      - REQ-134
      - REQ-156
      - REQ-157
      - REQ-158
      - REQ-166
      - REQ-167
      - REQ-188
      - REQ-194
      - REQ-198
      - REQ-211
      x-roles:
      - R09
      x-business-rules:
      - 개인정보 관리 원문 제약과 서버 권한을 적용한다
      - privacy_field_policies 저장 전 필수값·기간·중복 조건을 검증한다
      x-side-effects:
      - privacy_field_policies 레코드를 등록·수정하거나 실행 이력을 기록한다
      - audit_logs에 개인정보 관리 변경 전후 값을 기록한다
      x-required-tests:
      - 'happy: 유효한 개인정보 관리 저장 -> privacy_field_policies 후속 조회에 반영'
      - 'auth: R01이 개인정보 관리 저장 요청 -> 403 반환'
      - 'validation: 필수 식별자 누락 -> fields.id 오류'
      - 'business: 원문 불변/기간/중복 제약 위반 -> privacy_field_policies 상태 변경 없음'
      - 'side-effect: 성공 처리 후 audit_logs에 개인정보 관리 이력 존재'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiResponse'
                - type: object
                  properties:
                    error:
                      $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
    ApiResponse:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - true
        data: {}
        meta:
          type: object
    ApiError:
      type: object
      required:
      - success
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
        meta:
          type: object
...
security:
- sessionCookie: []
...